+++ src/_scss/pages/search/_collapsibleSidebar.scss
@@ -1,0 +1,158 @@
+@import '_searchFilter';
+
+$sidebar-width: 354px;
+$sidebar-width-closed: 32px;
+
+@keyframes collapsible-sidebar-slideopen {
+  0% {
+    width: $sidebar-width-closed;
+  }
+  100% {
+    width: $sidebar-width;
+  }
+}
+
+@keyframes collapsible-sidebar-slideclose {
+  0% {
+    width: $sidebar-width;
+  }
+  100% {
+    width: $sidebar-width-closed;
+  }
+}
+
+@keyframes drilldown-slidein {
+  0% {
+    display: block;
+    transform: translateX(24px);
+    visibility: hidden;
+  }
+  100% {
+    transform: translateX(0);
+    visibility: visible;
+  }
+}
+
+.search-collapsible-sidebar-container {
+  background-color: #fff;
+  box-shadow: unset;
+  border-radius: unset;
+  width: $sidebar-width;
+
+  .chevron {
+    margin: auto;
+    color: #2378c3;
+    width: 6px;
+    height: 10px;
+  }
+
+  .collapsible-sidebar--toggle {
+    float: right;
+    margin-right: -42px;
+    margin-top: 16px;
+    border-radius: 12px;
+    width: 24px;
+    height: 24px;
+    box-shadow: 0 1px 4px 0 rgba(0, 0, 0, 0.1);
+    border: solid 1px #dfe1e2;
+    background-color: #fff;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    cursor: pointer;
+  }
+
+  .collapsible-sidebar {
+    width: $sidebar-width;
+    margin-top: -40px !important; // TODO Keep here until we remove the legacy panel
+    border: 1px solid #dfe1d2;
+    box-shadow: none;
+    border-radius: unset;
+    padding: 0 rem(32);
+
+    &.opened:not(.is-initial-loaded) {
+      -webkit-animation: collapsible-sidebar-slideopen 0.25s forwards;
+      animation: collapsible-sidebar-slideopen 0.25s forwards;
+    }
+
+    &:not(.opened) {
+      animation: collapsible-sidebar-slideclose 0.25s forwards;
+      -webkit-animation: collapsible-sidebar-slideclose 0.25s forwards;
+
+      .collapsible-sidebar-drilldown, .collapsible-sidebar--main-menu {
+        overflow: hidden;
+        white-space: nowrap;
+      }
+    }
+
+    .collapsible-sidebar--header {
+      margin: rem(16) 0;
+      height: 24px;
+      font-size: 16px;
+      font-weight: 600;
+      line-height: 1.5;
+      color: #1b1b1b;
+    }
+
+    .sidebar-bottom-submit {
+      border-top: #dfe1e2 1px solid;
+
+    }
+
+    .collapsible-sidebar--main-menu {
+      &:not(.opened) {
+        display: none;
+      }
+    }
+
+    .collapsible-sidebar--drilldown {
+      &:not(.opened) {
+        display: none;
+      }
+
+      &.opened {
+        display: block;
+        -webkit-animation: drilldown-slidein 0.02s forwards;
+        animation: drilldown-slidein 0.02s forwards;
+      }
+
+      .collapsible-sidebar--back-btn {
+        color: #2378c3;
+        cursor: pointer;
+        .chevron {
+          width: 6px;
+          height: 10px;
+          margin: 2px 8px 2px 0;
+        }
+      }
+
+      .categories-list-item-container {
+        margin-top: 16px;
+        padding-bottom: 4px;
+        min-height: 26px;
+        display: flex;
+        flex-direction: row;
+        justify-content: flex-start;
+        line-height: 1.5;
+        align-items: center;
+        font-size: 12px;
+        color: #1b1b1b;
+        font-weight: 600;
+        border-bottom: solid 1px #dfe1e2;
+
+        .categories-list-item {
+          display: flex;
+
+        }
+      }
+
+      .search-filter__content {
+        margin: unset;
+      }
+      
+      .location-filter {
+        padding: 0 0 2rem 0;
+      }
+    }
+  }
+}
