+++ src/_scss/pages/search/_searchFilter.scss
@@ -5,7 +5,7 @@
   margin-bottom: $global-margin * 2;
 
   .search-filter__content {
-    margin: rem(16) rem(32);
+    margin: rem(16) 0;
 
     .search-filter__top-row {
       display: flex;
