--- src/_scss/pages/search/_sliderSidebar.scss
@@ -1,161 +1,0 @@
-@import '_searchFilter';
-
-$sidebar-width: 354px;
-$sidebar-width-closed: 32px;
-
-@keyframes left-sidebar-slideopen {
-  0% {
-    width: $sidebar-width-closed;
-  }
-  100% {
-    width: $sidebar-width;
-  }
-}
-
-@keyframes left-sidebar-slideclose {
-  0% {
-    width: $sidebar-width;
-  }
-  100% {
-    width: $sidebar-width-closed;
-  }
-}
-
-@keyframes drilldown-slidein {
-  0% {
-    transform: translateX(24px);
-    visibility: hidden;
-  }
-  100% {
-    transform: translateX(0);
-    visibility: visible;
-  }
-}
-
-
-
-.slider-open-toggle {
-  float: right;
-  margin-right: -28px;
-  margin-top: 16px;
-  border-radius: 12px;
-  width: 24px;
-  height: 24px;
-  box-shadow: 0 1px 4px 0 rgba(0, 0, 0, 0.1);
-  border: solid 1px #dfe1e2;
-  background-color: #fff;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  cursor: pointer;
-
-  .chevron {
-    margin: auto;
-    color: #2378c3;
-    width: 6px;
-    height: 10px;
-  }
-}
-
-.search-sidebar-slider {
-  background-color: #fff;
-  box-shadow: unset;
-  border-radius: unset;
-  width: $sidebar-width;
-
-  .slider-container {
-    width: $sidebar-width;
-    margin-top: -40px !important; // TODO Keep here until we remove the legacy panel
-    border: 1px solid #dfe1d2;
-    box-shadow: none;
-    border-radius: unset;
-    padding-right: 16px;
-
-    &.opened:not(.initial-load) {
-      -webkit-animation: left-sidebar-slideopen 0.25s forwards;
-      animation: left-sidebar-slideopen 0.25s forwards;
-    }
-
-    &:not(.opened) {
-      animation: left-sidebar-slideclose 0.25s forwards;
-      -webkit-animation: left-sidebar-slideclose 0.25s forwards;
-
-      .search-sidebar__drilldown, .search-sidebar__main-menu {
-        overflow: hidden;
-        white-space: nowrap;
-      }
-    }
-
-    .search-sidebar__drilldown {
-      margin: 0 32px 32px;
-
-      .categories-list-item-container {
-        margin-top: 16px;
-        padding-bottom: 4px;
-        min-height: 26px;
-        display: flex;
-        flex-direction: row;
-        justify-content: flex-start;
-        line-height: 1.5;
-        align-items: center;
-        font-size: 12px;
-        color: #1b1b1b;
-        font-weight: 600;
-        border-bottom: solid 1px #dfe1e2;
-
-        .categories-list-item {
-          display: flex;
-
-        }
-      }
-
-      .search-filter__content {
-        margin: unset;
-      }
-
-      .search-sidebar--header {
-        margin: 16px 0 16px 0;
-      }
-
-      .sidebar-back-btn {
-        color: #2378c3;
-        cursor: pointer;
-        .chevron {
-          width: 6px;
-          height: 10px;
-          margin: 2px 8px 2px 0;
-        }
-      }
-
-      .location-filter {
-        padding: 0 0 2rem 0;
-      }
-
-      &.opened {
-        -webkit-animation: drilldown-slidein 0.02s forwards;
-        //-webkit-animation-delay: 2s;
-        animation: drilldown-slidein 0.02s forwards;
-        //animation-delay: 2s;
-        display: block;
-      }
-    }
-
-    .search-sidebar__drilldown:not(.opened), .search-sidebar__main-menu:not(.opened) {
-      display: none;
-    }
-  }
-
-  .search-sidebar--header {
-    margin: rem(16) rem(32);
-    height: 24px;
-    font-size: 16px;
-    font-weight: 600;
-    line-height: 1.5;
-    color: #1b1b1b;
-  }
-
-  .sidebar-bottom-submit {
-    border-top: #dfe1e2 1px solid;
-
-  }
-}
