+++ src/_scss/pages/search/searchPage.scss
@@ -15,7 +15,7 @@
 
     @import "resultsView/resultsView";
     @import "resultsView/dtuiOverrides";
-    @import "./_sliderSidebar";
+    @import "collapsibleSidebar";
 
 
     .subaward-dropdown__list {
