+++ src/js/components/search/newSidebar/CategoryFilter.jsx
