+++ src/js/components/search/newSidebar/CateogriesList.jsx
